Add ReceiptListPage to pages and add Receipt component
